--- src/components/Header/Mobile/index.ts
@@ -1,3 +1,0 @@
-export { default as MobileMainHeader } from './MobileMainHeader';
-export { default as MobileDetailHeader } from './MobileDetailHeader';
-export { default as MobileEditorHeader } from './MobileEditorHeader';
